Refactor: Move SetupScreenScaffold to commonMain and update dependencies
