+++ composeApp/build.gradle.kts
@@ -35,9 +35,6 @@ kotlin {
             implementation(libs.koin.android)
             implementation(libs.koin.androidx.compose)
 
-            implementation("androidx.compose.material3.adaptive:adaptive:1.2.0-beta01")
-            implementation("androidx.compose.material3.adaptive:adaptive-layout:1.2.0-beta01")
-            implementation("androidx.compose.material3.adaptive:adaptive-navigation:1.2.0-beta01")
         }
         commonMain.dependencies {
             implementation(compose.runtime)
@@ -49,6 +46,7 @@ kotlin {
             implementation(libs.androidx.lifecycle.viewmodelCompose)
             implementation(libs.androidx.lifecycle.runtimeCompose)
 
+            implementation(libs.adaptive.z)
 
             api(libs.koin.core)
             implementation(libs.koin.compose)
