+++ gradle/libs.versions.toml
@@ -1,4 +1,5 @@
 [versions]
+adaptive_z = "1.2.0-alpha06"
 agp = "8.11.1"
 android-compileSdk = "36"
 android-minSdk = "24"
@@ -11,15 +12,16 @@ androidx-lifecycle = "2.9.3"
 androidx-testExt = "1.3.0"
 composeMultiplatform = "1.8.2"
 junit = "4.13.2"
-kotlin = "2.2.10"
+kotlin = "2.2.20"
 
-koin = "3.6.0-Beta4"
-koinComposeMultiplatform = "1.2.0-Beta4"
-navigationCompose = "2.8.0-alpha02"
-lifecycleViewModel = "2.8.2"
+koin = "4.1.1"
+koinComposeMultiplatform = "4.1.1"
+navigationCompose = "2.9.4"
+lifecycleViewModel = "2.9.4"
 
 
 [libraries]
+adaptive_z = { module = "org.jetbrains.compose.material3.adaptive:adaptive*", version.ref = "adaptive_z" }
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
 kotlin-testJunit = { module = "org.jetbrains.kotlin:kotlin-test-junit", version.ref = "kotlin" }
 junit = { module = "junit:junit", version.ref = "junit" }
